+++ src/shared/api/schemas/auth/refresh.ts
@@ -1,5 +1,8 @@
 import z from 'zod'
-import { ErrorResponseSchema, TokenResponseSchema } from './common'
+import {
+  ErrorResponseSchema,
+  TokenResponseSchema,
+} from '@/shared/api/schemas/auth/common'
 
 export const TokenRefreshRequestSchema = z.object({
   refresh_token: z.string().optional(),
